1. fix a bug in vector reverse 2. refactoring vector to make it easy to specialize in the future
