+++ jscomp/all.depend
@@ -1,6 +1,7 @@
 ext/bal_map.cmx : ext/bal_map_common.cmx ext/bal_map.cmi
 ext/bal_map_common.cmx :
-ext/bal_tree.cmx : ext/bal_tree.cmi
+ext/bal_set_common.cmx :
+ext/bal_tree.cmx : ext/bal_set_common.cmx ext/bal_tree.cmi
 ext/ext_array.cmx : ext/ext_array.cmi
 ext/ext_bytes.cmx : ext/ext_bytes.cmi
 ext/ext_char.cmx : ext/ext_char.cmi
@@ -10,6 +11,7 @@ ext/ext_filename.cmx : ext/literals.cmx ext/ext_string.cmx \
 ext/ext_format.cmx : ext/ext_format.cmi
 ext/ext_hashtbl.cmx : ext/ext_hashtbl.cmi
 ext/ext_ident.cmx : ext/string_set.cmx ext/ext_string.cmx ext/ext_ident.cmi
+ext/ext_int.cmx :
 ext/ext_io.cmx : ext/ext_pervasives.cmx ext/ext_io.cmi
 ext/ext_list.cmx : ext/ext_list.cmi
 ext/ext_map.cmx : ext/ext_map.cmi
@@ -24,7 +26,8 @@ ext/ext_scc.cmx : ext/int_vec_vec.cmx ext/int_vec.cmx ext/ext_scc.cmi
 ext/ext_string.cmx : ext/ext_bytes.cmx ext/ext_string.cmi
 ext/ext_sys.cmx : ext/ext_sys.cmi
 ext/hash_set.cmx : ext/hash_set.cmi
-ext/ident_set.cmx : ext/ext_format.cmx ext/bal_tree.cmx ext/ident_set.cmi
+ext/ident_set.cmx : ext/ext_format.cmx ext/bal_set_common.cmx \
+    ext/ident_set.cmi
 ext/int_map.cmx : ext/int_map.cmi
 ext/int_vec.cmx : ext/resize_array.cmx ext/int_vec.cmi
 ext/int_vec_vec.cmx : ext/resize_array.cmx ext/int_vec.cmx \
@@ -34,11 +37,11 @@ ext/ordered_hash_map.cmx : ext/ordered_hash_map.cmi
 ext/ordered_hash_set.cmx : ext/ordered_hash_set.cmi
 ext/resize_array.cmx : ext/ext_array.cmx ext/resize_array.cmi
 ext/string_map.cmx : ext/bal_map_common.cmx ext/string_map.cmi
-ext/string_set.cmx : ext/bal_tree.cmx ext/string_set.cmi
+ext/string_set.cmx : ext/bal_set_common.cmx ext/string_set.cmi
 ext/string_vec.cmx : ext/resize_array.cmx ext/string_vec.cmi
 ext/union_find.cmx : ext/union_find.cmi
 ext/bal_map.cmi : ext/bal_map_common.cmx
-ext/bal_tree.cmi :
+ext/bal_tree.cmi : ext/bal_set_common.cmx
 ext/ext_array.cmi :
 ext/ext_bytes.cmi :
 ext/ext_char.cmi :
@@ -550,7 +553,8 @@ ounit/oUnitTypes.cmx :
 ounit/oUnitUtils.cmx : ounit/oUnitTypes.cmx
 ounit_tests/ounit_array_tests.cmx : ounit/oUnit.cmx ext/ext_string.cmx \
     ext/ext_array.cmx
-ounit_tests/ounit_bal_tree_tests.cmx : ounit/oUnit.cmx ext/bal_tree.cmx
+ounit_tests/ounit_bal_tree_tests.cmx : ounit/oUnit.cmx ext/ext_int.cmx \
+    ext/bal_tree.cmx ext/bal_set_common.cmx
 ounit_tests/ounit_hash_set_tests.cmx : ext/ordered_hash_set.cmx \
     ounit/oUnit.cmx ext/hash_set.cmx
 ounit_tests/ounit_json_tests.cmx : ext/string_map.cmx ounit/oUnit.cmx \
