+++ jscomp/bin/whole_compiler.d
@@ -130,8 +130,7 @@ bin/whole_compiler.ml : ../ocaml/driver/compmisc.ml
 bin/whole_compiler.ml : ../ocaml/driver/compmisc.mli
 bin/whole_compiler.ml : common/ext_log.ml
 bin/whole_compiler.ml : common/ext_log.mli
-bin/whole_compiler.ml : ext/bal_tree.ml
-bin/whole_compiler.ml : ext/bal_tree.mli
+bin/whole_compiler.ml : ext/bal_set_common.ml
 bin/whole_compiler.ml : ext/string_set.ml
 bin/whole_compiler.ml : ext/string_set.mli
 bin/whole_compiler.ml : ext/ext_ident.ml
