+++ jscomp/ext/ident_set.ml
@@ -23,6 +23,7 @@
  * Foundation, Inc., 59 Temple Place - Suite 330, Boston, MA 02111-1307, USA. *)
 
 type elt = Ident.t
+
 let compare_elt (x : elt) (y : elt) = 
   let a =  Pervasives.compare (x.stamp : int) y.stamp in 
   if a <> 0 then a 
@@ -33,36 +34,48 @@ let compare_elt (x : elt) (y : elt) =
 
 
 (**************************************************************************************************)
-type nonrec t = elt Bal_tree.t 
-let empty = Bal_tree.empty 
-let is_empty = Bal_tree.is_empty
-let iter = Bal_tree.iter
-let fold = Bal_tree.fold
-let for_all = Bal_tree.for_all 
-let exists = Bal_tree.exists 
-let singleton = Bal_tree.singleton 
-let cardinal = Bal_tree.cardinal
-let elements = Bal_tree.elements
-let min_elt = Bal_tree.min_elt
-let max_elt = Bal_tree.max_elt
-let choose = Bal_tree.choose 
-let of_sorted_list = Bal_tree.of_sorted_list
-let of_sorted_array = Bal_tree.of_sorted_array
-let partition = Bal_tree.partition 
-let filter = Bal_tree.filter 
-let of_sorted_list = Bal_tree.of_sorted_list
-let of_sorted_array = Bal_tree.of_sorted_array
+type nonrec t = elt Bal_set_common.t 
+let empty = Bal_set_common.empty 
+let is_empty = Bal_set_common.is_empty
+let iter = Bal_set_common.iter
+let fold = Bal_set_common.fold
+let for_all = Bal_set_common.for_all 
+let exists = Bal_set_common.exists 
+let singleton = Bal_set_common.singleton 
+let cardinal = Bal_set_common.cardinal
+let elements = Bal_set_common.elements
+let min_elt = Bal_set_common.min_elt
+let max_elt = Bal_set_common.max_elt
+let choose = Bal_set_common.choose 
+let of_sorted_list = Bal_set_common.of_sorted_list
+let of_sorted_array = Bal_set_common.of_sorted_array
+let partition = Bal_set_common.partition 
+let filter = Bal_set_common.filter 
+let of_sorted_list = Bal_set_common.of_sorted_list
+let of_sorted_array = Bal_set_common.of_sorted_array
 
-let rec add x (tree : _ Bal_tree.t) : _ Bal_tree.t =
+let rec split x (tree : t) : t *  bool * t  =
+  match tree with 
+  | Empty ->
+    (Empty, false, Empty)
+  | Node(l, v, r, _) ->
+    let c = compare_elt x v in
+    if c = 0 then (l, true, r)
+    else if c < 0 then
+      let (ll, pres, rl) = split x l in (ll, pres, Bal_set_common.internal_join rl v r)
+    else
+      let (lr, pres, rr) = split x r in (Bal_set_common.internal_join l v lr, pres, rr)
+
+let rec add x (tree : _ Bal_set_common.t) : _ Bal_set_common.t =
   match tree with  
   | Empty -> Node(Empty, x, Empty, 1)
   | Node(l, v, r, _) as t ->
     let c = compare_elt x v in
     if c = 0 then t else
-    if c < 0 then Bal_tree.internal_bal (add x l) v r else Bal_tree.internal_bal l v (add x r) 
+    if c < 0 then Bal_set_common.internal_bal (add x l) v r else Bal_set_common.internal_bal l v (add x r) 
 
 
-let rec mem x (tree : _ Bal_tree.t) = 
+let rec mem x (tree : _ Bal_set_common.t) = 
   match tree with 
   | Empty -> false
   | Node(l, v, r, _) ->
@@ -84,8 +97,8 @@ let rec remove x (tree : t ) : t =
   | Empty -> Empty
   | Node(l, v, r, _) ->
     let c = compare_elt x v in
-    if c = 0 then Bal_tree.internal_merge l r else
-    if c < 0 then Bal_tree.internal_bal (remove x l) v r else Bal_tree.internal_bal l v (remove x r)
+    if c = 0 then Bal_set_common.internal_merge l r else
+    if c < 0 then Bal_set_common.internal_bal (remove x l) v r else Bal_set_common.internal_bal l v (remove x r)
 
 
 
@@ -96,13 +109,13 @@ let rec union (s1 : t ) (s2 : t ) =
   | (Node(l1, v1, r1, h1), Node(l2, v2, r2, h2)) ->
     if h1 >= h2 then
       if h2 = 1 then add v2 s1 else begin
-        let (l2, _, r2) = Bal_tree.split v1 s2 in
-        Bal_tree.internal_join (union l1 l2) v1 (union r1 r2)
+        let (l2, _, r2) = split v1 s2 in
+        Bal_set_common.internal_join (union l1 l2) v1 (union r1 r2)
       end
     else
     if h1 = 1 then add v1 s2 else begin
-      let (l1, _, r1) = Bal_tree.split v2 s1 in
-      Bal_tree.internal_join (union l1 l2) v2 (union r1 r2)
+      let (l1, _, r1) = split v2 s1 in
+      Bal_set_common.internal_join (union l1 l2) v2 (union r1 r2)
     end    
 
 let rec inter (s1 : t) (s2 : t) : t =
@@ -110,11 +123,11 @@ let rec inter (s1 : t) (s2 : t) : t =
   | (Empty, t2) -> Empty
   | (t1, Empty) -> Empty
   | (Node(l1, v1, r1, _), t2) ->
-    begin match Bal_tree.split v1 t2 with
+    begin match split v1 t2 with
       | (l2, false, r2) ->
-        Bal_tree.internal_concat (inter l1 l2) (inter r1 r2)
+        Bal_set_common.internal_concat (inter l1 l2) (inter r1 r2)
       | (l2, true, r2) ->
-        Bal_tree.internal_join (inter l1 l2) v1 (inter r1 r2)
+        Bal_set_common.internal_join (inter l1 l2) v1 (inter r1 r2)
     end 
 
 let rec diff (s1 : t) (s2 : t) : t =
@@ -122,11 +135,11 @@ let rec diff (s1 : t) (s2 : t) : t =
   | (Empty, t2) -> Empty
   | (t1, Empty) -> t1
   | (Node(l1, v1, r1, _), t2) ->
-    begin match Bal_tree.split v1 t2 with
+    begin match split v1 t2 with
       | (l2, false, r2) ->
-        Bal_tree.internal_join (diff l1 l2) v1 (diff r1 r2)
+        Bal_set_common.internal_join (diff l1 l2) v1 (diff r1 r2)
       | (l2, true, r2) ->
-        Bal_tree.internal_concat (diff l1 l2) (diff r1 r2)    
+        Bal_set_common.internal_concat (diff l1 l2) (diff r1 r2)    
     end
 
 (**************************************************************************************************)
