+++ jscomp/test/bal_set_common.ml
@@ -1,0 +1,1 @@
+../ext/bal_set_common.ml
